+++ src/Tenanti/Migrator/Factory.php
@@ -110,7 +110,8 @@ class Factory implements FactoryInterface
      */
     public function runInstall(Model $entity, $database)
     {
-        $table = $this->resolveMigrationTableName($entity);
+        $database = $this->bindWithKey($entity, $database);
+        $table    = $this->resolveMigrationTableName($entity);
 
         $repository = $this->resolveMigrator($table)->getRepository();
 
@@ -131,6 +132,7 @@ class Factory implements FactoryInterface
      */
     public function runUp(Model $entity, $database, $pretend = false)
     {
+        $database = $this->bindWithKey($entity, $database);
         $table    = $this->resolveMigrationTableName($entity);
         $migrator = $this->resolveMigrator($table);
 
@@ -149,7 +151,8 @@ class Factory implements FactoryInterface
      */
     public function runDown(Model $entity, $database, $pretend = false)
     {
-        $table = $this->resolveMigrationTableName($entity);
+        $database = $this->bindWithKey($entity, $database);
+        $table    = $this->resolveMigrationTableName($entity);
         $migrator = $this->resolveMigrator($table);
 
         $migrator->setConnection($database);
@@ -167,7 +170,8 @@ class Factory implements FactoryInterface
      */
     public function runReset(Model $entity, $database, $pretend = false)
     {
-        $table = $this->resolveMigrationTableName($entity);
+        $database = $this->bindWithKey($entity, $database);
+        $table    = $this->resolveMigrationTableName($entity);
         $migrator = $this->resolveMigrator($table);
 
         $migrator->setConnection($database);