+++ tests/Migrator/FactoryTest.php
@@ -144,6 +144,30 @@ class FactoryTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
+     * Test Orchestra\Tenanti\Migrator\Factory::runInstall()
+     * method when given custom attribute.
+     *
+     * @test
+     */
+    public function testRunInstallMethodWithCustomAttribute()
+    {
+        $app    = $this->getAppContainer();
+        $driver = 'user';
+        $config = array('migration' => 'migrations');
+
+        $app['schema']->shouldReceive('hasTable')->once()->with('migrations')->andReturn(false)
+            ->shouldReceive('create')->once()->with('migrations', m::type('Closure'))->andReturnNull();
+
+        $app['db']->shouldReceive('connection')->with('tenant_foo')->andReturnSelf()
+            ->shouldReceive('getSchemaBuilder')->andReturn($app['schema']);
+
+        $stub  = new Factory($app, $driver, $config);
+        $model = $this->getMockModel();
+
+        $this->assertNull($stub->runInstall($model, 'tenant_{entity.username}'));
+    }
+
+    /**
      * Test Orchestra\Tenanti\Migrator\Factory::runUp()
      * method.
      *
@@ -235,7 +259,8 @@ class FactoryTest extends \PHPUnit_Framework_TestCase
     {
         $model = m::mock('\Illuminate\Database\Eloquent\Model');
 
-        $model->shouldReceive('getKey')->andReturn(5);
+        $model->shouldReceive('getKey')->andReturn(5)
+            ->shouldReceive('toArray')->once()->andReturn(array('name' => 'Administrator', 'username' => 'foo'));
 
         return $model;
     }