+++ docs/changes.md
@@ -9,6 +9,7 @@ title: Tenanti Change Log
 
 * Allow to specifically declare database connection name when running `Orchestra\Tenanti\Observer`. This allow user to setup multi-tenancy on a different database.
 * Allow to configuration custom multi-tenant migration table name.
+* Bind tenant key to database name, which allow using multi-tenancy database using multiple database connection.
 
 ### v2.2.0 {#v2-2-0}